requirements_chatbot.txt
requests==2.32.3
pandas==2.2.3
numpy<2.0
matplotlib==3.10.1
scikit-learn==1.6.1
flask==3.1.0
beautifulsoup4==4.13.3
pytest==8.3.5
mlx
mlx-lm
tqdm
psutil
python-multipart
fastapi==0.104.1
uvicorn==0.24.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
torch==2.1.2
rank_bm25==0.2.2
anthropic==0.25.1
requests==2.32.3